Web Version

+++ Scripts/ActiveBush.cs
@@ -18,7 +18,7 @@ public partial class ActiveBush : Area2D
 	}
 
 	// Called every frame. 'delta' is the elapsed time since the previous frame.
-	public override void _Process(double delta)
+	public override void _Process(float delta)
 	{
 	}
